Talos: Re-organize 'App' components together
Now that a lot of the redundant state has been cleared up,
the structure appears more clearly, and the initial move
seems to have been slightly erroneous.
