--- talos/src/components/App.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import Preview from "./preview/preview"
-import StepWrapper from "./steps/stepWrapper"
-
-import './app.scss'
-
-const App = () => {
-  return (
-    <div>
-      <Preview />
-
-      <StepWrapper />
-    </div>
-  )
-}
-
-export default App;
