+++ talos/src/components/app/app.scss
@@ -1,5 +1,5 @@
-@import "./../base.scss";
-@import "./../grid.scss";
+@import "../../base.scss";
+@import "../../grid.scss";
 
 * {
   @include box-sizing(border-box);
