+++ talos/src/components/items/controls/buttonToPrevStep.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { withAppState } from "../../steps/AppState"
+import { withAppState } from "../../app/State"
 
 import ButtonReset from "./buttonReset"
