+++ talos/src/components/preview/buttonToPreviewLink.js
@@ -4,7 +4,7 @@
  */
 import React from 'react';
 
-import { withAppState } from "../steps/AppState"
+import { withAppState } from "../app/State"
 
 import ButtonReset from "../items/controls/buttonReset"
