+++ talos/src/components/request/requestDialog.js
@@ -5,8 +5,8 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import { withStyles } from '@material-ui/core/styles';
 
-import { REQUEST } from "../steps/AppAction"
-import { withAppState } from "../steps/AppState"
+import { REQUEST } from "../app/Action"
+import { withAppState } from "../app/State"
 
 import Loader from "../items/static/loader"
 import PrevButton from "../items/static/prevButton"
