+++ talos/src/components/steps/desktopMenuList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { steps } from "./static"
 
-import { withAppState } from "./AppState"
+import { withAppState } from "../app/State"
 
 import ButtonStep from "../items/controls/buttonStep"
